better_duck_sys

Function duckdb_appender_create_query

pub unsafe extern "C" fn duckdb_appender_create_query(
    connection: duckdb_connection,
    query: *const c_char,
    column_count: idx_t,
    types: *mut duckdb_logical_type,
    table_name: *const c_char,
    column_names: *mut *const c_char,
    out_appender: *mut duckdb_appender,
) -> duckdb_state

Creates an appender object that executes the given query with any data appended to it.

Note that the object must be destroyed with duckdb_appender_destroy.

@param connection The connection context to create the appender in. @param query The query to execute, can be an INSERT, DELETE, UPDATE or MERGE INTO statement. @param column_count The number of columns to append. @param types The types of the columns to append. @param table_name (optionally) the table name used to refer to the appended data, defaults to “appended_data”. @param column_names (optionally) the list of column names, defaults to “col1”, “col2”, … @param out_appender The resulting appender object. @return DuckDBSuccess on success or DuckDBError on failure.