better_duck_sys

Function duckdb_execute_n_tasks_state

pub unsafe extern "C" fn duckdb_execute_n_tasks_state(
    state: duckdb_task_state,
    max_tasks: idx_t,
) -> idx_t

Execute DuckDB tasks on this thread.

The thread will keep on executing tasks until either duckdb_finish_execution is called on the state, max_tasks tasks have been executed or there are no more tasks to be executed.

Multiple threads can share the same duckdb_task_state.

@param state The task state of the executor @param max_tasks The maximum amount of tasks to execute @return The amount of tasks that have actually been executed