better_duck_sys

Function duckdb_execute_pending

pub unsafe extern "C" fn duckdb_execute_pending(
    pending_result: duckdb_pending_result,
    out_result: *mut duckdb_result,
) -> duckdb_state

Fully execute a pending query result, returning the final query result.

If duckdb_pending_execute_task has been called until DUCKDB_PENDING_RESULT_READY was returned, this will return fast. Otherwise, all remaining tasks must be executed first.

Note that the result must be freed with duckdb_destroy_result.

@param pending_result The pending result to execute. @param out_result The result object. @return DuckDBSuccess on success or DuckDBError on failure.