better_duck_sys

Function duckdb_from_time_tz

pub unsafe extern "C" fn duckdb_from_time_tz(
    micros: duckdb_time_tz,
) -> duckdb_time_tz_struct

Decompose a TIME_TZ objects into micros and a timezone offset.

Use duckdb_from_time to further decompose the micros into hour, minute, second and microsecond.

@param micros The time object, as obtained from a DUCKDB_TYPE_TIME_TZ column.